# api_server/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: info@payloadops.com
    name: Payload Ops Team
  description: "The Payload Ops API streamlines management of AI prompts, projects,\
    \ organizations, teams, and users through conventional HTTP requests. This platform\
    \ enables efficient automation and control of resources, providing robust tools\
    \ for developers to manage settings, memberships, and activities seamlessly."
  title: Payload Ops API
  version: "1.0"
servers:
- description: Production server (uses live data)
  url: https://api.payloadops.com/v1
- description: Sandbox server (uses test data)
  url: https://sandbox-api.payloadops.com/v1
security:
- ApiKeyAuth: []
- BearerAuth: []
paths:
...
  /users:
    get:
      description: |
        Retrieves a comprehensive list of all users registered in the system. This operation is essential for administrators and systems that need to manage user access, monitor user activity, or gather user data for reporting purposes.

        **Key Features**:
        - **Broad Overview**: Provides a detailed view of all user profiles, including their roles, status, and basic information.
        - **User Management Support**: Facilitates the administration and oversight of user accounts, supporting tasks such as auditing, role assignment, and access reviews.

        **Use Cases**:
        - System administrators needing to monitor user activities or manage user permissions.
        - Automated systems that integrate user data for analytics or marketing purposes.
      operationId: listUsers
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'
                type: array
          description: "Successfully retrieved a list of user profiles, each including\
            \ essential details such as names, email addresses, and roles."
        "500":
          description: "A server error occurred, preventing the listing of users."
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - Users
    post:
      description: |
        Creates a new user with specified details such as name, email, and role. This endpoint is crucial for adding new users to the system, facilitating controlled access to applications and services based on predefined roles and permissions.

        **Key Features**:
        - **Flexible User Creation**: Allows for the dynamic creation of user accounts with configurable roles and permissions.
        - **Security and Compliance**: Adheres to security best practices by ensuring that all new users are created with appropriate access levels and data validation.

        **Use Cases**:
        - Onboarding new employees by creating user accounts as part of organizational access provisioning.
        - Registering new customers or clients on platforms requiring user authentication and personalization.

        **Requirements**:
        - Input data must include essential user details such as name, email, and a predefined role. Additional data such as department or group might also be required based on system configuration.
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
        description: "Contains the necessary attributes for creating a new user, including\
          \ name, email, role, and any additional relevant information."
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: "User created successfully, with the user's profile details\
            \ returned including identifiers and status."
        "400":
          description: "Invalid input, typically due to missing required fields, non-conformity\
            \ to data formats, or invalid values."
        "500":
          description: "A server error occurred, preventing the creation of a new\
            \ user."
      security:
      - BearerAuth: []
      summary: Create a new user
      tags:
      - Users
  /users/{userId}:
    delete:
      description: |
        Permanently deletes a user from the system based on their unique ID. This operation is crucial for managing user lifecycle, particularly in cases where users need to be removed due to security concerns or upon user request.

        **Key Features**:
        - **Secure User Management**: Ensures that user data can be permanently removed in compliance with data protection regulations.
        - **Lifecycle Management**: Supports the management of user accounts by facilitating the complete removal of user profiles when necessary.

        **Use Cases**:
        - Administrators needing to remove user access permanently, either at the user's request or due to policy requirements.
        - Managing user deactivation and data deletion in compliance with privacy laws and regulations.
      operationId: deleteUser
      parameters:
      - description: The unique identifier of the user to be deleted.
        explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: "User deleted successfully, no content returned."
        "404":
          description: No user found with the specified ID.
        "500":
          description: "A server error occurred, preventing the deletion of the user."
      security:
      - BearerAuth: []
      summary: Delete a specific user
      tags:
      - Users
    get:
      description: |
        Retrieves detailed information about a user, identified by their unique ID. This operation is essential for system administrators and applications requiring user data validation, profile management, or access rights verification.

        **Key Features**:
        - **Detailed User Profiles**: Provides comprehensive details about the user, including contact information, roles, and permissions.
        - **Security and Compliance**: Supports security processes by enabling identity verification and user management.

        **Use Cases**:
        - Administrators managing user access and permissions within a system.
        - Services integrating user information for personalization or configuration.
      operationId: getUser
      parameters:
      - description: The unique identifier of the user to retrieve.
        explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: Successfully retrieved detailed information about the user.
        "404":
          description: No user found with the specified ID.
        "500":
          description: "A server error occurred, preventing the retrieval of user\
            \ information."
      security:
      - BearerAuth: []
      summary: Get a specific user by ID
      tags:
      - Users
    put:
      description: |
        Updates the profile or settings of a user identified by their unique ID. This endpoint allows for the modification of user details such as names, contact information, or roles, supporting ongoing management of user data.

        **Key Features**:
        - **Profile Updates**: Enables changes to user profiles, supporting dynamic updates to contact details, preferences, and roles.
        - **Access Control Adjustments**: Facilitates updates to user permissions and access controls, essential for maintaining security and operational integrity.

        **Use Cases**:
        - Administrators updating user roles or permissions in response to organizational changes.
        - Users updating their own profiles as part of regular account maintenance.
      operationId: updateUser
      parameters:
      - description: The unique identifier of the user to update.
        explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
        description: Contains the updated data for the user's profile.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: User information updated successfully.
        "400":
          description: "Invalid input, such as malformed data or incomplete required\
            \ fields."
        "404":
          description: No user found with the specified ID.
        "500":
          description: "A server error occurred, preventing the update of the user."
      security:
      - BearerAuth: []
      summary: Update a specific user
      tags:
      - Users
...
  /projects/{projectId}/keys/{keyId}:
    delete:
      description: |
        Deletes a specific API key from a project, permanently removing its access to the project resources. This operation is essential for maintaining security when an API key is no longer needed or if its permissions need to be revoked urgently.

        **Key Features**:
        - **Secure Cleanup**: Ensures that unused or compromised API keys are removed to prevent unauthorized access.
        - **Access Revocation**: Provides a method to immediately invalidate an API key when necessary.

        **Use Cases**:
        - Removing access for team members who no longer require it, such as after project completion or role changes.
        - Revoking keys that may have been exposed or compromised to maintain security integrity.
      operationId: deleteApiKey
      parameters:
      - description: The unique identifier of the project from which the API key will
          be deleted.
        explode: false
        in: path
        name: projectId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the API key to be deleted.
        explode: false
        in: path
        name: keyId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: "The API key was deleted successfully, no content returned."
        "404":
          description: Either the API key or the project was not found.
        "500":
          description: "A server error occurred, preventing the deletion of the API\
            \ key."
      security:
      - BearerAuth: []
      summary: Delete a specific API key
      tags:
      - API Keys
    get:
      description: |
        Retrieves detailed information about a specific API key associated with a project. This operation is crucial for managing access to project resources, allowing administrators to review the permissions and details of individual API keys.

        **Key Features**:
        - **Security Management**: Provides the ability to monitor and manage API access through detailed key information.
        - **Access Review**: Enables project managers to verify the scope and validity of an API key, ensuring that it aligns with current security policies.

        **Use Cases**:
        - Administrators auditing access permissions for compliance and security best practices.
        - Developers checking API key details to troubleshoot access issues or to ensure proper configuration.
      operationId: getApiKey
      parameters:
      - description: The unique identifier of the project associated with the API
          key.
        explode: false
        in: path
        name: projectId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the API key to retrieve.
        explode: false
        in: path
        name: keyId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
          description: Detailed information about the API key retrieved successfully.
        "404":
          description: Either the specified API key or the project was not found.
        "500":
          description: "A server error occurred, preventing the retrieval of the API\
            \ key."
      security:
      - BearerAuth: []
      summary: Retrieve a specific API key
      tags:
      - API Keys
    put:
      description: |
        Updates the scopes associated with a specific API key. This operation allows administrators to modify access levels and permissions for API keys, adapting to changing security needs or project requirements.

        **Key Features**:
        - **Scope Management**: Adjusts the access boundaries defined by the API key scopes, ensuring they are kept up-to-date with project roles and permissions.
        - **Flexible Access Control**: Offers the ability to dynamically change API key permissions without needing to recreate keys.

        **Use Cases**:
        - Modifying API key permissions in response to role changes within a project team.
        - Updating key scopes to restrict or expand access as part of security policy adjustments.
      operationId: updateApiKey
      parameters:
      - description: The unique identifier of the project for which the API key is
          managed.
        explode: false
        in: path
        name: projectId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the API key to be updated.
        explode: false
        in: path
        name: keyId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyInput'
        description: JSON payload containing the new scopes for the API key.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
          description: The API key's scopes were updated successfully.
        "400":
          description: "Invalid input, such as unspecified or unsupported scopes."
        "404":
          description: Either the API key or the project was not found.
        "500":
          description: "A server error occurred, preventing the update of the API\
            \ key."
      security:
      - BearerAuth: []
      summary: Update an API key's scopes
      tags:
      - API Keys
...
components:
  schemas:
...
    User:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        name: name
        id: id
        email: email
        updatedAt: updatedAt
      properties:
        id:
          description: Unique identifier for the user
          type: string
        name:
          description: User's full name (optional)
          type: string
        email:
          description: User's email address
          type: string
        createdAt:
          description: Timestamp of user creation
          format: date-time
          type: string
        updatedAt:
          type: string
      type: object
    UserInput:
      example:
        name: name
        email: email
      properties:
        name:
          description: User's full name (optional)
          type: string
        email:
          description: User's email address
          type: string
      required:
      - email
      type: object
...
    ApiKey:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        id: id
        scopes:
        - scopes
        - scopes
        projectId: projectId
        key: key
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          description: Unique identifier for the API key
          type: string
        key:
          description: The API key token
          type: string
        scopes:
          description: List of scopes or permissions granted by this API key
          items:
            type: string
          type: array
        projectId:
          description: The project ID this API key belongs to
          type: string
        createdAt:
          description: Timestamp when the API key was created
          format: date-time
          type: string
        updatedAt:
          description: Timestamp when the API key was last updated
          format: date-time
          type: string
      type: object
    ApiKeyInput:
      example:
        scopes:
        - scopes
        - scopes
      properties:
        scopes:
          description: List of scopes or permissions to be granted to this API key
          items:
            type: string
          type: array
      required:
      - scopes
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      description: |
        API Key authentication provides a straightforward method for clients to authenticate requests. This scheme expects an API key to be supplied in the headers of each request. Clients must include the API key in the 'X-API-KEY' header field.

        **How to use**:
        - Obtain an API key from the API provider.
        - Include the API key in the header of each request to the API as follows:
          ```
          X-API-KEY: {api_key}
          ```
        - The API validates the key against its database, and if the key is valid, allows access to the requested resources.

        **Advantages**:
        - Simple and easy to implement.
        - Suitable for accessing public APIs where security requirements are not excessively strict.

        **Limitations**:
        - Less secure than more complex authentication methods, such as tokens, as API keys can be easily exposed in transit or storage.
        - API keys do not expire by default and revoking them requires direct intervention.
      in: header
      name: X-API-KEY
      type: apiKey
    BearerAuth:
      bearerFormat: JWT
      description: |
        Bearer Token authentication uses JSON Web Tokens (JWT) to verify the identity of the requester. This scheme is included in the Authorization header of each HTTP request using the Bearer format.

        **How to use**:
        - Obtain a JWT from the authentication server. Typically, this involves an initial request with credentials (username/password) to receive the token.
        - Include the token in the Authorization header of your HTTP requests as follows:
          ```
          Authorization: Bearer {token}
          ```
        - The server decodes the JWT to verify its validity and authorizes the request based on the token's payload and signature.

        **Advantages**:
        - Offers more security than API key authentication as the token content is protected by cryptographic algorithms.
        - Tokens can be designed to expire after a certain period, enhancing security by limiting the duration of their validity.

        **Limitations**:
...